Implement solitaire logic
This commit allows you to play a full game of solitaire. Drawing from
the stock, taking from the waste, and putting cards to the foundation
are all implemented.

Tests are lacking for the solitaire logic - this is because I'm pretty
sure I'm going to change to the Sawayama Solitaire rules (from
Zachtronic's Last Call BBS).